fileApi.js created
this is an api service for sending and downloading files from appropriate endpoints as marked in the api documentation.

all files are uploaded to the appropriate endpoints and can be downloaded using the file hash-id.

uploadPostFiles - uploads files to the post endpoint
uploadCommentFiles - uploads files to the comment endpoint
uploadQuestionFiles - uploads files to the question endpoint
uploadTestFiles - uploads files to the test endpoint
uploadBoardBannerFiles - uploads files to the board banner endpoint
uploadDescBannerFiles - uploads files to the desc banner endpoint
uploadUserAvatarFiles - uploads files to the user avatar endpoint
uploadUserBannerFiles - uploads files to the user banner endpoint
downloadFile - downloads files using the file hash-id